(chore) extract functionality to useActiveExercise; add exercise_id to key to enforce uniqueness
